% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obisidian.R
\name{clean_text_content}
\alias{clean_text_content}
\title{Clean text content}
\usage{
clean_text_content(text)
}
\arguments{
\item{text}{Character vector to clean}
}
\value{
Cleaned text
}
\description{
Clean text content
}
\keyword{internal}
