% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obisidian.R
\name{format_lit_notes_to_json}
\alias{format_lit_notes_to_json}
\title{Format Literature Notes to JSON}
\usage{
format_lit_notes_to_json(
  lit_notes,
  note_separator_pattern = "\\\\s*\\\\*\\\\*Note:\\\\*\\\\*\\\\s*",
  include_only_commented = FALSE
)
}
\arguments{
\item{lit_notes}{List of literature notes, or a tibble/data.frame as returned by \code{\link{get_all_obsidian_notes_content}}.}

\item{note_separator_pattern}{Regex pattern to separate annotation text from its note.}

\item{include_only_commented}{Logical, if TRUE, only commented annotations are included. Defaults to FALSE.}
}
\value{
JSON formatted string
}
\description{
Format Literature Notes to JSON
}
