% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{initialize_chat}
\alias{initialize_chat}
\title{Initialize LLM Chat}
\usage{
initialize_chat(
  model = "gpt-4-0125-preview",
  system_prompt = NULL,
  api_args = list()
)
}
\arguments{
\item{system_prompt}{System context}

\item{api_args}{Additional API arguments}

\item{model_type}{Type of model ("claude", "openai", "gemini")}
}
\value{
Chat object
}
\description{
Initialize LLM Chat
}
