% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\docType{data}
\name{template_configs}
\alias{template_configs}
\title{Template configurations for Obsidian notes}
\format{
An object of class \code{list} of length 1.
}
\usage{
template_configs
}
\description{
Predefined and user-defined template configurations for extracting
content from different Obsidian note templates
}
\keyword{datasets}
